+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html);
 however, insignificant breaking changes do not guarantee a major version bump, see the reasoning [here](https://github.com/kyb3r/modmail/issues/319). If you're a plugin developer, note the "BREAKING" section.
 
+# v3.10.3
+
+### Improved
+
+- Thread genesis message now shows other recipients.
+
+### Fixed
+
+- `?snippet add` now properly blocks command names.
+
+### Internal
+
+- Set `LOG_DISCORD` environment variable to the logger level and log discord events.
+
 # v3.10.2
 This is a hotfix for react to contact.